requirements.txt
numpy==2.2.1
opencv_python==4.10.0.84
pillow==10.4.0
pillow_heif==0.21.0
pytesseract==0.3.13
qreader==3.14
surya-ocr==0.13.1
tensorboard==2.19.0
torchmetrics==1.7.0
